Only spent 10 minutes on this and with the brain not firing on all cylinders. But printing to the console seems to show that you have an infinite loop running.
The issue is that the for loop on line 71 is based on the length of the triangles array. However, with every iteration of this loop we grow that triangles array. The result is that you never get off that block of code.

I discovered this by simply logging the length of the triangles array on line 75. As you can see if you run locally it gets over 100,000 pretty fast. I expect it eventually crashes from memory loss.

I'm not sure why this was not happening with lower iterations. Probably has something to do with your logic on line 72.

At any rate, I hope that gets you started in the right direction. I will try to take another look after work tomorrow.

(Side note: Python is super picky about variable types. So if you are printing an integer concatenated with a string you will need to convert the number to a string type using str(). See lines 69 and 75. No other language I've worked with makes you do this.)